[#1375] Revise public API
move ProductsResource and Factory to internal package
